+++ CHANGELOG.md
@@ -6,9 +6,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Upgraded Sql Server library from `System.Data.SqlClient` to `Microsoft.Data.SqlClient`
+
 ### Added
 
 - Added `CreateDatabaseTimeoutInSeconds` static property to `DiscoveredServerHelper`
+
 
 ## [1.0.7] - 2021-05-18